Fixed a few more leaks and did some refactoring
